IrregularDelimitedFiles

Documentation for IrregularDelimitedFiles.

IrregularDelimitedFiles.gen_data

IrregularDelimitedFiles.read_as_dict

IrregularDelimitedFiles.read_irregular

IrregularDelimitedFiles.gen_data — Method

gen_data(filename; options...)

Writes a file with num_rows rows of data of the form

key#value1,value2,…,valueN

The number of values for a given row - N in the example - is generated randomly and is at most max_num_values.

Inputs

filename: Where to write the data.

num_rows::Int = 3: Number of rows of data.

value_range::UnitRange = 1:1_000_000: Range of values.

max_num_values::Int = 10: Maximum number of values per row (key not included).

value_distribution_type::Symbol = :uniform: How to generate the number of values per row.

:uniform: Uniform distribution taken from 1:max_num_values.

:exponential: Exponentially distributed value. Take a number α ∈ [0,log(max_num_values)] and return $⌊exp(α)⌋$

key_sep::Char = '#': Key separator

array_sep::Char = ',': Array separator

IrregularDelimitedFiles.read_as_dict — Method

output = read_as_dict(filename; options...)

Read irregular file as a dictionary of integers to vector of integers.

Inputs

filename: File name.

key_sep::Char = '#': Character separating the key from the values.

array_sep::Char = ',': Character separating values.

IrregularDelimitedFiles.read_irregular — Method

keys, arrays, values = read_irregular(filename; options...)

Read an irregular file like the one below:

1#1024
271#31,32,33,301
403#99,105

and returns three equal-sized arrays with this information such that they can be used to create a table. For instance, for the file above, it would return:

keys = [ 1, 271, 271, 271, 271, 403, 403]

indexes = [ 0, 0, 1, 2, 3, 0, 1]

values = [1024, 31, 32, 33, 301, 99, 105]

Input

filename: File name.

key_sep::Char = '#': Character separating the key from the values.

array_sep::Char = ',': Character separating values.

Output

Vectors keys::Vector{Int}, indexes::Vector{Int} and values::Vector{Int} such that a table provides the same information. See the example above.